+++ src/main/java/com/alkemy/disney/service/impl/MovieServiceImpl.java
@@ -28,24 +28,27 @@ public class MovieServiceImpl implements MovieService {
 
     ModelMapper modelMapper = new ModelMapper();
 
-    public List<MovieListDTO> getAllMovies(){
+    public List<MovieListDTO> getAllMovies() {
         List<MovieEntity> moviesEntities = movieRepository.findAll();
         List<MovieListDTO> moviesDTO = new ArrayList<>();
-        for (MovieEntity m:moviesEntities) {
+        for (MovieEntity m : moviesEntities) {
             MovieListDTO mDTO = modelMapper.map(m, MovieListDTO.class);
             moviesDTO.add(mDTO);
         }
         return moviesDTO;
     }
 
-    public MovieDetailDTO getMovieById(Long id){
-            MovieEntity movieEntity = movieRepository.getMovieById(id);
-            MovieDetailDTO movieDTO = movieMapper.movieEntityToDTO(movieEntity);
-            return movieDTO;
+    public MovieDetailDTO getMovieById(Long id) {
+        Optional<MovieEntity> movieEntity = movieRepository.findById(id);
+        if (!movieEntity.isPresent()) {
+            throw new RuntimeException("Movie with id " + id + " not found");
+        }
+        MovieDetailDTO movieDTO = movieMapper.movieEntityToDTO(movieEntity.get());
+        return movieDTO;
 
     }
 
-    public MovieDetailDTO saveMovie(MovieDetailDTO movie){
+    public MovieDetailDTO saveMovie(MovieDetailDTO movie) {
         MovieEntity movieEntity = movieMapper.movieDTOToEntity(movie);
         movieEntity.setCreationDate(LocalDate.now());
         MovieEntity savedMovie = movieRepository.save(movieEntity);
@@ -56,22 +59,24 @@ public class MovieServiceImpl implements MovieService {
     @Override
     public void deleteMovie(Long id) {
         Optional<MovieEntity> movie = movieRepository.findById(id);
+        if (!movie.isPresent()) {
+            throw new RuntimeException("Movie with id " + id + " not found");
+        }
         movieRepository.deleteById(id);
     }
 
     @Override
     public MovieDetailDTO updateMovie(Long id, MovieUpdateDTO movieDTO) {
-            GenreEntity genre = modelMapper.map(movieDTO.getGenre(), GenreEntity.class);
-            MovieEntity movie = modelMapper.map(movieDTO, MovieEntity.class);
 
-            Optional<MovieEntity> movieEntity = movieRepository.findById(id);
-            MovieEntity movieToUpdate = movieEntity.get();
-            movieToUpdate.setTitle(movie.getTitle());
-            movieToUpdate.setImage(movie.getImage());
-            movieToUpdate.setScore(movie.getScore());
-            movieToUpdate.setGenre(genre);
 
-            MovieEntity updatedMovie = movieRepository.save(movieToUpdate);
-            return movieMapper.movieEntityToDTO(updatedMovie);
+        Optional<MovieEntity> movieEntity = movieRepository.findById(id);
+
+        if (!movieEntity.isPresent()) {
+            throw new RuntimeException("Movie with id " + id + " not found");
+        }
+        MovieEntity movieToUpdate = movieEntity.get();
+
+        MovieEntity updatedMovie = movieRepository.save(movieMapper.updateMovieMapper(movieDTO, movieToUpdate));
+        return movieMapper.movieEntityToDTO(updatedMovie);
     }
 }
